+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Documentation for Paper objects.
 - Path constructor typechecking input Vector array.
+- Installed ESLint with eslint-prettier-config.
 
 ### Fixed